Image possibilities
I am trying to figure out how some of these might work. I tried to cut away and make "blendy" the .png as much as I could. The inverse space situation is a bit special and could be awesome or could be a nightmare of ugly. We are going to have to go back and forth on this a bit I think to see what works and what is an epic fail.